Move test to `test_basics`
